\name{DDPNA-package}
\alias{DDPNA-package}
\alias{DDPNA}
\docType{package}
\title{
Disease-drived Differential Proteins And Proteomic Co-expression Network Associated Analysis
}
\description{
disease drived proteins associated network in different species crosstalk.
The package is used to analysis differential proteomics consensus network in two or more datasets.
The function \code{Data_impute} need \pkg{impute} package from Bioconductor, the function \code{ID_match} and the function \code{MaxQdataconvert} need \pkg{Biostrings} package from Bioconductor.
}
\details{
\tabular{ll}{
Package: \tab DDPNA\cr
Type: \tab Package\cr
Version: \tab 0.3.3\cr
Creat Data: \tab 2019-03-18\cr
Date: \tab 2024-03-13\cr
License: \tab GPL (>= 2)\cr
}
This package is used for construct core component sets associated co-expression network.
It offered a preprocess function and core component sets subnetwork construct.
Preprocess function can process proteomics data included data extract, outlier removal and impute. Some functions are fit for other omics data.
}
\author{
Kefu Liu

Maintainer: Kefu Liu <liukefu19@163.com>
}


\keyword{ package }

