% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Dforimpute}
\alias{Dforimpute}
\title{a data used for Dataimpute function.}
\format{A list containing proteomic protein information and quantification data.}
\usage{
Dforimpute
}
\description{
a data used for Dataimpute function.
}
\keyword{internal}
