% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imputedData}
\alias{imputedData}
\title{an imputed data used for downstream analysis}
\format{A list containing proteomic protein information and quantification data.}
\usage{
imputedData
}
\description{
an imputed data used for downstream analysis
}
\keyword{internal}
