% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{net}
\alias{net}
\title{A network which is obtained by WGCNA-package blockwiseModule function.
it can short the time of Module analysis example time.}
\format{A list containing network module information, ME information and dendrograms information.}
\usage{
net
}
\description{
A network which is obtained by WGCNA-package blockwiseModule function.
it can short the time of Module analysis example time.
}
\keyword{internal}
