% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ProteomicData}
\alias{ProteomicData}
\title{Small example proteomic quantification data extract from txt or csv file used to
demonstrate data extract function.}
\format{A list containing 2 data frame: one is MaxQ proteomic data; another is other type of proteomic data}
\usage{
ProteomicData
}
\description{
Small example proteomic quantification data extract from txt or csv file used to
demonstrate data extract function.
}
\keyword{internal}
