% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sample_ID_data}
\alias{Sample_ID_data}
\title{A small uniprot ID information used to do ID convert in example.}
\format{A data frame corresponding uniprot ID information}
\usage{
Sample_ID_data
}
\description{
A small uniprot ID information used to do ID convert in example.
}
\keyword{internal}
