\name{DDPNA-package}
\alias{DDPNA-package}
\alias{DDPNA}
\docType{package}
\title{
Disease-drived Differential Proteins And Proteomic Co-expression Network Associated Analysis
}
\description{
disease drived proteins associated network in different species crosstalk.
The package is used to analysis differential proteomics consensus network in two or more datasets.
The function \code{Data_impute} need \pkg{impute} package from Bioconductor, the function \code{ID_match} and the function \code{MaxQdataconvert} need \pkg{Biostrings} package from Bioconductor.
}
\details{
\tabular{ll}{
Package: \tab DHMDA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2019-03-18\cr
License: \tab GPL (>= 2)\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Kefu Liu

Maintainer: Kefu Liu <lkf1013@gmail.com>
}


\keyword{ package }

